internal/filetransfer: split out FTP tests from file_transfer_test.go
