+++ internal/filetransfer/ftp_test.go
@@ -1,0 +1,283 @@
+// Copyright 2019 The Moov Authors
+// Use of this source code is governed by an Apache License
+// license that can be found in the LICENSE file.
+
+package filetransfer
+
+import (
+	"bytes"
+	"errors"
+	"fmt"
+	"io/ioutil"
+	"math/rand"
+	"os"
+	"path/filepath"
+	"strings"
+	"testing"
+	"time"
+
+	"github.com/moov-io/base"
+	mhttptest "github.com/moov-io/paygate/internal/httptest"
+
+	filedriver "github.com/goftp/file-driver"
+	"github.com/goftp/server"
+	"github.com/jlaffaye/ftp"
+)
+
+var (
+	portSource = rand.NewSource(time.Now().Unix())
+)
+
+func port() int {
+	return int(30000 + (portSource.Int63() % 9999))
+}
+
+func createTestFTPServer(t *testing.T) (*server.Server, error) {
+	t.Helper()
+	if testing.Short() {
+		t.Skip("skipping due to -short")
+	}
+	opts := &server.ServerOpts{
+		Auth: &server.SimpleAuth{
+			Name:     "moov",
+			Password: "password",
+		},
+		Factory: &filedriver.FileDriverFactory{
+			RootPath: filepath.Join("..", "..", "testdata", "ftp-server"),
+			Perm:     server.NewSimplePerm("test", "test"),
+		},
+		Hostname: "localhost",
+		Port:     port(),
+		Logger:   &server.DiscardLogger{},
+	}
+	svc := server.NewServer(opts)
+	if svc == nil {
+		return nil, errors.New("nil FTP server")
+	}
+	if err := try(func() error { return svc.ListenAndServe() }, 50*time.Millisecond); err != nil {
+		if err == errTimeout {
+			return svc, nil
+		}
+		return nil, err
+	}
+	return svc, nil
+}
+
+var errTimeout = errors.New("timeout exceeded")
+
+// try will attempt to call f, but only for as long as t. If the function is still
+// processing after t has elapsed then errTimeout will be returned.
+func try(f func() error, t time.Duration) error {
+	answer := make(chan error)
+	go func() {
+		answer <- f()
+	}()
+	select {
+	case err := <-answer:
+		return err
+	case <-time.After(t):
+		return errTimeout
+	}
+}
+
+func TestTry(t *testing.T) {
+	start := time.Now()
+
+	err := try(func() error {
+		time.Sleep(50 * time.Millisecond)
+		return nil
+	}, 1*time.Second)
+
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	diff := time.Since(start)
+
+	if diff < 50*time.Millisecond {
+		t.Errorf("%v was under 50ms", diff)
+	}
+	if limit := 2 * 100 * time.Millisecond; diff > limit {
+		t.Errorf("%v was over %v", diff, limit)
+	}
+}
+
+func createTestFTPConnection(t *testing.T, svc *server.Server) (*ftp.ServerConn, error) {
+	t.Helper()
+	conn, err := ftp.DialTimeout(fmt.Sprintf("localhost:%d", svc.Port), 10*time.Second)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if err := conn.Login("moov", "password"); err != nil {
+		t.Fatal(err)
+	}
+	return conn, nil
+}
+
+func TestFTP(t *testing.T) {
+	svc, err := createTestFTPServer(t)
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer svc.Shutdown()
+
+	conn, err := createTestFTPConnection(t, svc)
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer conn.Quit()
+
+	dir, err := conn.CurrentDir()
+	if err != nil {
+		t.Fatal(err)
+	}
+	if dir == "" {
+		t.Error("empty current dir?!")
+	}
+
+	// Change directory
+	if err := conn.ChangeDir("scratch"); err != nil {
+		t.Error(err)
+	}
+
+	// Read a file we know should exist
+	resp, err := conn.RetrFrom("existing-file", 0) // offset of 0
+	if err != nil {
+		t.Error(err)
+	}
+	bs, _ := ioutil.ReadAll(resp)
+	bs = bytes.TrimSpace(bs)
+	if !bytes.Equal(bs, []byte("Hello, World!")) {
+		t.Errorf("got %q", string(bs))
+	}
+}
+
+func TestFTP__tlsDialOption(t *testing.T) {
+	if testing.Short() {
+		return // skip network calls
+	}
+
+	cafile, err := mhttptest.GrabConnectionCertificates(t, "google.com:443")
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer os.Remove(cafile)
+
+	opt, err := tlsDialOption(cafile)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if opt == nil {
+		t.Fatal("nil tls DialOption")
+	}
+}
+
+func TestFTP__getInboundFiles(t *testing.T) {
+	svc, agent := createTestFileTransferAgent(t)
+	defer agent.Close()
+	defer svc.Shutdown()
+
+	files, err := agent.GetInboundFiles()
+	if err != nil {
+		t.Fatal(err)
+	}
+	if len(files) != 1 {
+		t.Errorf("got %d files", len(files))
+	}
+	if files[0].Filename != "iat-credit.ach" {
+		t.Errorf("files[0]=%s", files[0])
+	}
+	bs, _ := ioutil.ReadAll(files[0].Contents)
+	bs = bytes.TrimSpace(bs)
+	if !strings.HasPrefix(string(bs), "101 121042882 2313801041812180000A094101Bank                   My Bank Name                   ") {
+		t.Errorf("got %v", string(bs))
+	}
+
+	// make sure we perform the same call and get the same result
+	files, err = agent.GetInboundFiles()
+	if err != nil {
+		t.Fatal(err)
+	}
+	if len(files) != 1 {
+		t.Errorf("got %d files", len(files))
+	}
+	if files[0].Filename != "iat-credit.ach" {
+		t.Errorf("files[0]=%s", files[0])
+	}
+}
+
+func TestFTP__getReturnFiles(t *testing.T) {
+	svc, agent := createTestFileTransferAgent(t)
+	defer agent.Close()
+	defer svc.Shutdown()
+
+	files, err := agent.GetReturnFiles()
+	if err != nil {
+		t.Fatal(err)
+	}
+	if len(files) != 1 {
+		t.Errorf("got %d files", len(files))
+	}
+	if files[0].Filename != "return-WEB.ach" {
+		t.Errorf("files[0]=%s", files[0])
+	}
+	bs, _ := ioutil.ReadAll(files[0].Contents)
+	bs = bytes.TrimSpace(bs)
+	if !strings.HasPrefix(string(bs), "101 091400606 6910001341810170306A094101FIRST BANK & TRUST     ASF APPLICATION SUPERVI        ") {
+		t.Errorf("got %v", string(bs))
+	}
+
+	// make sure we perform the same call and get the same result
+	files, err = agent.GetReturnFiles()
+	if err != nil {
+		t.Fatal(err)
+	}
+	if len(files) != 1 {
+		t.Errorf("got %d files", len(files))
+	}
+	if files[0].Filename != "return-WEB.ach" {
+		t.Errorf("files[0]=%s", files[0])
+	}
+}
+
+func TestFTP__uploadFile(t *testing.T) {
+	svc, agent := createTestFileTransferAgent(t)
+	defer agent.Close()
+	defer svc.Shutdown()
+
+	content := base.ID()
+	f := File{
+		Filename: base.ID(),
+		Contents: ioutil.NopCloser(strings.NewReader(content)), // random file contents
+	}
+
+	// Create outbound directory
+	parent := filepath.Join("..", "..", "testdata", "ftp-server", agent.OutboundPath())
+	os.Mkdir(parent, 0777)
+
+	if err := agent.UploadFile(f); err != nil {
+		t.Fatal(err)
+	}
+
+	ftpAgent, _ := agent.(*FTPTransferAgent)
+
+	// manually read file contents
+	ftpAgent.conn.ChangeDir(agent.OutboundPath())
+	resp, _ := ftpAgent.conn.Retr(f.Filename)
+	if resp == nil {
+		t.Fatal("nil File response")
+	}
+	r, _ := ftpAgent.readResponse(resp)
+	if r == nil {
+		t.Fatal("failed to read file")
+	}
+	bs, _ := ioutil.ReadAll(r)
+	if !bytes.Equal(bs, []byte(content)) {
+		t.Errorf("got %q", string(bs))
+	}
+
+	// delete the file
+	if err := agent.Delete(f.Filename); err != nil {
+		t.Fatal(err)
+	}
+}
